moved the data to handling to the file containing the component
