+++ imports/ui/pages/home.jsx
@@ -1,10 +1,16 @@
 // NPM
 import React from 'react';
 
+// Meteor
+import { Meteor } from 'meteor/meteor';
+
+// Atmosphere
+import { createContainer } from 'meteor/react-meteor-data';
+
 // App
 import MeetupComponent from '/imports/ui/components/meetups.jsx';
-import ProjectContainer from '/imports/ui/containers/project.js';
-
+import { ProjectComponentContainer } from '/imports/ui/components/projects.jsx';
+import { Meetups } from '/imports/api/meetups/collections.js';
 
 export const HomePage = React.createClass({
     displayName: 'HomePage',
@@ -17,7 +23,7 @@ export const HomePage = React.createClass({
 
     renderProjects() {
         if (this.props.currentMeetup) {
-            return <ProjectContainer params={{ meetup: this.props.currentMeetup }} />
+            return <ProjectComponentContainer params={{ meetup: this.props.currentMeetup }} />
         }
     },
 
@@ -36,3 +42,18 @@ export const HomePage = React.createClass({
         );
     }
 });
+
+export const HomeContainer = createContainer(() => {
+    const handler = Meteor.subscribe("currentMeetup");
+    const data = {
+        isLoading: !handler.ready(),
+    };
+
+    if (!data.isLoading) {
+        const today = new Date().setHours(0);
+        data.currentMeetup = Meetups.findOne({
+            time: { $gt: today }
+        }, { limit: 1 });
+    }
+    return data;
+}, HomePage);
